backend/requirements.txt
# RepoGuard AI — Python Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.0
websockets==14.1

# Git operations
gitpython==3.1.43

# Static analysis
semgrep>=1.90.0
bandit>=1.8.0

# AI providers
groq==0.12.0
google-generativeai==0.8.0

# PDF generation
reportlab==4.2.4

# GitHub API
PyGithub==2.5.0

# Database
supabase==2.10.0

# Rate limiting
slowapi==0.1.9

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.2
